+++ Kernel/System/Crypt/PGP.pm
@@ -11,10 +11,13 @@ package Kernel::System::Crypt::PGP;
 use strict;
 use warnings;
 
+use Kernel::System::VariableCheck qw(:all);
+
 our @ObjectDependencies = (
     'Kernel::Config',
     'Kernel::System::Encode',
     'Kernel::System::FileTemp',
+    'Kernel::System::CheckItem',
     'Kernel::System::Log',
     'Kernel::System::Main',
     'Kernel::System::Time',
@@ -99,7 +102,6 @@ crypt a message
 sub Crypt {
     my ( $Self, %Param ) = @_;
 
-    # check needed stuff
     for my $ParamName (qw( Message Key )) {
         if ( !$Param{$ParamName} ) {
             $Kernel::OM->Get('Kernel::System::Log')->Log(
@@ -150,7 +152,6 @@ The returned hash %Result has the following keys:
 sub Decrypt {
     my ( $Self, %Param ) = @_;
 
-    # check needed stuff
     for (qw(Message)) {
         if ( !defined( $Param{$_} ) ) {
             $Kernel::OM->Get('Kernel::System::Log')->Log(
@@ -204,7 +205,6 @@ sign a message
 sub Sign {
     my ( $Self, %Param ) = @_;
 
-    # check needed stuff
     for (qw(Message Key)) {
         if ( !$Param{$_} ) {
             $Kernel::OM->Get('Kernel::System::Log')->Log(
@@ -241,8 +241,20 @@ sub Sign {
     my ( $FHPhrase, $FilePhrase ) = $FileTempObject->TempFile();
     print $FHPhrase $Pw;
     close $FHPhrase;
-    my $GPGOptions = qq{--passphrase-fd 0 --default-key $Param{Key} -o $FileSign $SigType $Filename};
-    my $LogMessage = qx{$Self->{GPGBin} $GPGOptions <$FilePhrase 2>&1};
+
+    my $Quiet = '';
+
+    # GnuPG 2.1 (and higher) may send info messages about used default keys to STDERR, which leads to problems.
+    if (
+        IsHashRefWithData( $Self->{Version} )
+        && sprintf( "%.3d%.3d", $Self->{Version}->{Major}, $Self->{Version}->{Minor} ) >= 2_001
+        )
+    {
+        $Quiet = '--quiet --batch --pinentry-mode=loopback';
+    }
+
+    my $GPGOptions = qq{$Quiet --passphrase-fd 0 -o $FileSign --default-key $Param{Key} $SigType $Filename};
+    my $LogMessage = qx{$Self->{GPGBin} $GPGOptions < $FilePhrase 2>&1};
 
     # error
     if ($LogMessage) {
@@ -293,7 +305,6 @@ The returned hash %Result has the following keys:
 sub Verify {
     my ( $Self, %Param ) = @_;
 
-    # check needed stuff
     if ( !$Param{Message} ) {
         $Kernel::OM->Get('Kernel::System::Log')->Log(
             Priority => 'error',
@@ -721,8 +732,32 @@ sub SecretKeyGet {
 
     my $Key = quotemeta( $Param{Key} || '' );
 
-    my $LogMessage = qx{$Self->{GPGBin} --export-secret-keys --armor $Key 2>&1};
-    my $SecretKey  = '';
+    my $LogMessage = '';
+
+    # GnuPG 2.1 (and higher) asks via pinentry for the key passphrase. We suppress that behavior by passing the phrase
+    # via STDIN from temporary file (--passphrase-fd 0 / file descriptor 0).
+    if (
+        IsHashRefWithData( $Self->{Version} )
+        && sprintf( "%.3d%.3d", $Self->{Version}->{Major}, $Self->{Version}->{Minor} ) >= 2_001
+        )
+    {
+        my %PasswordHash = %{ $Kernel::OM->Get('Kernel::Config')->Get('PGP::Key::Password') };
+        my $Password = $PasswordHash{$Key} || '';
+
+        my ( $FH, $Filename ) = $Kernel::OM->Get('Kernel::System::FileTemp')->TempFile();
+        print $FH $Password;
+        close $FH;
+
+        $LogMessage
+            = qx{$Self->{GPGBin} --batch --pinentry-mode=loopback --export-secret-keys --passphrase-fd 0 --armor --decrypt <$Filename 2>&1};
+    }
+
+    # GnuPG 2.0 (and lower)
+    else {
+        $LogMessage = qx{$Self->{GPGBin} --export-secret-keys --armor $Key 2>&1};
+    }
+
+    my $SecretKey = '';
 
     if ( $LogMessage =~ /nothing exported/i ) {
         $LogMessage =~ s/\n//g;
@@ -760,7 +795,6 @@ remove public key from key ring
 sub PublicKeyDelete {
     my ( $Self, %Param ) = @_;
 
-    # check needed stuff
     if ( !$Param{Key} ) {
         $Kernel::OM->Get('Kernel::System::Log')->Log(
             Priority => 'error',
@@ -800,7 +834,6 @@ remove secret key from key ring
 sub SecretKeyDelete {
     my ( $Self, %Param ) = @_;
 
-    # check needed stuff
     if ( !$Param{Key} ) {
         $Kernel::OM->Get('Kernel::System::Log')->Log(
             Priority => 'error',
@@ -856,7 +889,6 @@ add key to key ring
 sub KeyAdd {
     my ( $Self, %Param ) = @_;
 
-    # check needed stuff
     if ( !$Param{Key} ) {
         $Kernel::OM->Get('Kernel::System::Log')->Log(
             Priority => 'error',
@@ -907,13 +939,28 @@ sub _Init {
         $Self->{GPGBin} = "LC_MESSAGES=POSIX $Self->{GPGBin} $Self->{Options}";
     }
 
+    # determine active GnuPG version
+    my $VersionString = '';
+
+    eval {
+        $VersionString = `$Self->{GPGBin} --version`;
+    };
+
+    if ( $VersionString =~ m{ gpg [ ]+ \(.+?\) [ ]+ (\d+)\.(\d+)\.(\d+) }smx ) {
+        $Self->{Version} = {
+            Major  => $1,
+            Minor  => $2,
+            Patch  => $3,
+            String => $1 . '.' . $2 . '.' . $3,
+        };
+    }
+
     return $Self;
 }
 
 sub _DecryptPart {
     my ( $Self, %Param ) = @_;
 
-    # check needed stuff
     for (qw(Key Password Filename)) {
         if ( !defined( $Param{$_} ) ) {
             $Kernel::OM->Get('Kernel::System::Log')->Log(
@@ -924,16 +971,36 @@ sub _DecryptPart {
         }
     }
 
-    # get temp file object
     my $FileTempObject = $Kernel::OM->Get('Kernel::System::FileTemp');
 
+    # temp file for decrypted message
     my ( $FHDecrypt, $FileDecrypt ) = $FileTempObject->TempFile();
     close $FHDecrypt;
+
+    # temp file for passphrase
     my ( $FHPhrase, $FilePhrase ) = $FileTempObject->TempFile();
     print $FHPhrase $Param{Password};
     close $FHPhrase;
-    my $GPGOptions = qq{--batch --passphrase-fd 0 --yes --decrypt -o $FileDecrypt $Param{Filename}};
-    my $LogMessage = qx{$Self->{GPGBin} $GPGOptions <$FilePhrase 2>&1};
+
+    my $LogMessage = '';
+
+    # GnuPG 2.1 (and higher)
+    if (
+        IsHashRefWithData( $Self->{Version} )
+        && sprintf( "%.3d%.3d", $Self->{Version}->{Major}, $Self->{Version}->{Minor} ) >= 2_001
+        )
+    {
+        my $GPGOptions
+            = qq{--batch --pinentry-mode=loopback --passphrase-fd 0 --armor -o $FileDecrypt --decrypt $Param{Filename}};
+        $LogMessage = qx{$Self->{GPGBin} $GPGOptions < $FilePhrase 2>&1};
+    }
+
+    # GnuPG 2.0 (and lower)
+    else {
+        my $GPGOptions = qq{--batch --passphrase-fd 0 --yes --decrypt -o $FileDecrypt $Param{Filename}};
+        $LogMessage = qx{$Self->{GPGBin} $GPGOptions <$FilePhrase 2>&1};
+    }
+
     if ( $LogMessage =~ /failed/i ) {
         $Kernel::OM->Get('Kernel::System::Log')->Log(
             Priority => 'notice',
@@ -968,7 +1035,6 @@ Clean and build the log
 sub _HandleLog {
     my ( $Self, %Param ) = @_;
 
-    # check needed stuff
     for (qw(LogString)) {
         if ( !defined( $Param{$_} ) ) {
             $Kernel::OM->Get('Kernel::System::Log')->Log(
@@ -1110,7 +1176,11 @@ sub _ParseGPGKeyList {
             # only use last 8 chars of key-ID in order to be compatible with previous parser
             $Key{KeyPrivate} = substr( $Fields[4], -8, 8 );
         }
-        elsif ( $Type eq 'fpr' ) {
+
+        # The public and secret key information will be exposed at first. Since GnugPG 2 (and higher)
+        # the sub-key fingerprint will be displayed as well (--list-keys --with-colons), which leads
+        # to overwriting of the main information. Therefore we just collect the first fingerprint.
+        elsif ( $Type eq 'fpr' && !$Key{Fingerprint} && !$Key{Fingerprint} ) {
             $Key{FingerprintShort} = $Fields[9];
 
             # add fingerprint in standard format, too
@@ -1152,7 +1222,6 @@ sub _ParseGPGKeyList {
 sub _CryptedWithKey {
     my ( $Self, %Param ) = @_;
 
-    # check needed stuff
     if ( !$Param{File} ) {
         $Kernel::OM->Get('Kernel::System::Log')->Log(
             Priority => 'error',